Remove cool off period for submitted Apply Again applications
The cool off period to allow references to be received and candidates to
make edits doesn't make a lot of sense in Apply Again, as it is likely
that the candidate already has their references and they have just
edited their application. The delay could instead disadvantage them as
they can only apply to one course at a time.

When submitting an application, check to see if it's Apply Again and if
enough references have been provided. If so, set its application choice
edit_by to the current time (no cool off period) and ensure its status
is 'application_complete'.