Remove cool off period when receiving Apply Again references
An Apply Again application may have removed previous references and
added new ones. When the minimum number of required references are
received (currently 2), we want to send the application to the Provider
as soon as possible rather than wait out the cool off period as we do
during a standard application.

- SubmitReference already progresses the state of application choices if
  enough references have been received. Add to this logic by checking if
  the application form is apply_again, and if so set the
  application_choice edit_by to the current time.
- Restructure the SubmitReference spec for clarity and add an example
  for the Apply Again logic.