Remove .uniq from site factory
While it is true, that there is a constraint in the Find system that
stops providers from creating sites past an upper limit because
otherwise the characters would overlap, it doesn't seem necessary to
replicate this in our factories.

Names don't need to be unique as far as I know.