Move logic to not render providers if less than 2 to page state
This was in the filter_component which didn't make much sense, as it
only applies to providers an not other filters like
accredited_providers.

If there are less than two providers then there is not need for a filter
as it would be the same as not having a provider filter, as all
applications would belong to that filter.

For other filters this is not the case.

When the logic was in the filter_component it was applyed to more than
just providers, which was not correct.

This commit also makes the page_state spec a bit more robust by using
more realistic test subjects and adding a few more, specific,
expectations.