Make permissions changes human-readable in the audit log
Before this change, assigning a provider user to a provider resulted in
an audit log entry like "Create ProviderPermissions #24, provider_id:
9".

This is hopeless for seeing what actually happened, so render a better
string. For creation and deletion events the provider_id is present in
the audited_changes hash. For update events the id is not necessarily
present, and in the event the permission was later revoked, the original
Provider record might not exist. In this case, look up the original
creation record to discover the associated provider_id.