Add spec for application ordering.
Adds spec to ensure that application cards render in desc 'last updated'
order in the provider_interface/application_choices/index.erb.html.

NOTE: this is really a reintroduction given that there was a spec with
this name removed a few months ago when sorting was removed from this
index page.

Introducing a specific test now as it is a requirement that the page
loads applications in this way.