Add filter toggle
The future commits will will make the filter only appear on small
screens. As we use a basic headless browser (Rack test) there will be no
automated tests with this piece of functionality. As agreed by the dev
team a tech debt ticket will be added to look at this again in the
round.