Make URL value for filters_toggle reflect state of page
The URL was false when the filter was visible and true when it wasn't
the page 'worked' but this was the oposite value to what would be
expected so this commit reverses that.

i.e. if filters_visible is true the filter is show if false then it is
not.