Add initial HESA dataset
The Higher Education Statistics Agency provides structured data to
describe the following characteristics of degree courses:

- classes (grades)
- establishments
- subjects
- types

We want to use this data when candidates specify their degree as part of
their applications.

Using raw CSV files as a source, add an initial cut of this data in the
form of several arrays so that we can start using it in code.