Fix Withdrawn by us and Rejected filters
'Withdrawn by us' filter was showing no results and 'Rejected' filter included
both rejected and withdrawn by us applications. This was caused by the fact
that the status filter was running WHERE status query. However
`offer_withdrawn` does not exist in the database; those applications have
`rejected` status in the database.

To fix it we separately add offer_withdrawn_choices when this filter is
selected.  When the 'Rejected' filter is selected, we need to exclude
'Withdrawn by us' applications from the results.