Add application reference to vendor API responses
One of the vendors pointed out that it would be useful to have
application_reference since it’s displayed to the candidates and providers in
the UI. Application reference is often used when candidates contact providers.
Adding it to the API will allow provider's system to display this information
and therefore to identify candidates/application by the reference.