Show a sensible error when API clients try to repeat offers
An API user may

- POST :id/offer to make an offer on the default course
- POST :id/offer again without course information

At the moment the app will 404 on the second request as the
ProviderAuthorisation code will try to look up the course option for a
"ChangeOffer" (as opposed to a "MakeOffer") request. When there is a nil
course option, it raises RecordNotFound - we then rescue that in the
controller and ascribe it to the ApplicationChoice itself not being
found, which yields a weird and unhelpful error for the client.

Clients might be wrong to do this, but it's confusing and we should do
better.

Fix it by falling back to the application_choice.course_option in the
event that no course option is supplied for a "ChangeOffer"
request.