Add filter on small screens
NOTE: this functionality uses the MoJ Frontend which has a dependency on
jQuery. Rather than rebuilding parts of the MoJ Frontend ourselves and
then having to support it, our proposed approach is to merge this with jQuery
then upstream a fix into moj-frontend to take out the jQuery dependency
for this particular widget, then take it out of our app too.

This commit adds the filter in a progressive way to devices with small
screens.

When JS enabled:

- on large screens the filter is present and the 'hide/show filter' button is not present.

- on small screens the the filter collapses to a flaoting box on the side and can be
hidden and shown.

Whe JS is not enabled:

- on large screens the filter is present on the left.

- on small screens the filter is moved above the application cards.