Allow TestApplications to remain application_complete with completed references
When we submit the last reference,
GetApplicationChoicesReadyToSendToProvider (via
SendApplicationsToProvider)
compares the edit_by of the application with the current date. As we're
in a fake timeline in the past here, edit_by is always in the past, so
we always send the choice and we never end up in the :application_complete state.

Cheat by setting edit_by before we receive references.