Specify explicit sort order on degree review page
The candidate_entering_degrees_spec is failing intermittently on Azure.
The spec assumes an order when multiple degrees are present on the page.
If they are out of order, the wrong degree gets deleted and subsequent
assertions about the presence of the other degree fail.

It may be worth investigating why implicit ordering by primary key isn't
occurring on Azure test runs, but in the meantime let's specify an
explicit order when we render the review page.