SyncProviderFromFind depends on provider_recruitment_cycle_year
We can't rely on the global recruitment_cycle_year any more, so pass it
in when we sync a provider. We can't pass the whole provider in as it
might not exist yet. (We could probably throw together a struct, but
that's more or less what we're doing in the method call anyway).

Specs are updated to pass the currently-stubbed recruitment_cycle_year.