Return early from initDegreeTypeAutocomplete Javascript
If the feature flag 'hesa_degree_data' is switched off, we don't render
the container that this function is dependent on. This leads to a
Javascript error in the browser when we try to retrieve the source data
embedded on the container element.

To fix this, move the check for the existence of the container to an
earlier point in the function, prior to the data retrieval.