Change referee email template from html to text
The reference_request_chase_again_email is currently an html template.
This appears to be a mistake, as all other emails are sent as plain
text. Change the template extension to correct this.

This was noticed while looking at an intermittent failure in the
RefereeMailer spec. We use Faker to generate referee names. An
apostrophe in the surname gets escaped when sending the email as HTML.
This led to failures when asserting for the presence of the name in the
email body.