Account for duplicates which may exist for RatifyingProviderPermissions
Records created by the sync from Find may have used the new STI type 'RatifyingProviderPermissions' for new records
 these will have the same foreign keys as existing records using the STI type 'ProviderInterface::AccreditedBodyPermissions'
so attempting to update them to the use the new STI type will fail on the [type, ratifying_provider_id, training_provider_id]
uniqueness constraint. This causes the migration to fail.
This commit checks for the existence of a newer record (using the RatifyingProviderPermissions type) and either disgards the old
record or updates it depending on whether a 'duplicate' exists.