Merge "provider_add_provider_users" and "provider_make_decisions_restriction" feature flags
We currently have 3 feature flags that interact with each other. The
`provider_add_provider_users` flag allows providers to add other
providers. `provider_make_decisions_restriction` allows the providers
to then restrict some users from making decisions (the third
`enforce_provider_to_provider_permissions` then extends this to
provider-to-provider permissions, which we're not touching at the
moment).

This PR merges the 2 feature flags, because we're going to launch the
features at the same time. It also makes it easier to understand the
code.