Remove cucumber and tests
We introduced cucumber at the start of the build to increase our
understanding and communicate how things should work. Over time we've
built more acceptance tests and built a lot of documentation. The
cucumber tests aren't actively maintained anymore, and the tests are
covered by other specs.

This removes the tests and the cuke infrastructure itself.