Add functionality to persist IELTS qualification details
Implement the 'English as a foreign language' flow up to the point of
being able to submit and have persisted details of an IELTS
qualification.

The IeltsForm gathers all required attributes and does a wholesale
replace of the candidate's EnglishLanguageProficiency record. This
ensures all database state changes required by this part of the EFL
flow happen in one place, reducing the likelihood of unexpected state
errors.