The wizard knows how to persist itself
Given a hash (which in practice will be the session) the wizard can save
and rehydrate itself from that hash. This is the final step to decouple
the wizard from HTTP.

It's not practical to always save_state! when we instantate the Wizard,
nice as that would be. If the user is checking answers and they enter
invalid input, they could obliterate their old answer.