Add functionality to persist TOEFL qualification details
Enhance the 'English as a foreign language' flow to allow the candidate
to provide details about their TOEFL qualification.

The ToeflForm gathers all required attributes and does a wholesale
replace of the candidate's EnglishLanguageProficiency record. This
ensures all database state changes required by this part of the EFL
flow happen in one place, reducing the likelihood of unexpected state
errors.