Remove all code associated with Provider relationship management
We're preparing to refactor the underlying models, and large swathes of
this code will be rewritten to come into line with the prototype. We
will keep it in history, but in the meantime take it out so that we
don't spend ages fixing tests that we'll need to rewrite anyway.

Once the refactor is done we can reincorporate this code.