Add a PhaseBanner component and always display it with the header
This will change for the provider header, so pull it up into a
component (which is faster than a partial anyway).

The provider landing page now has a phase banner. This means there is no
unsightly gap around the coloured bar under the header, so we can lose
the negative margin that brought it flush with the black header.