Hide Languages section of Personal Details when EFL is active
We're adding an 'English as a foreign language' section to the
application form, visible to any candidate with a nationality other than
British and Irish. This has a lot of overlap with the existing Languages
page of the Personal Details section, so we should hide it if the EFL
feature is active.

- Add an object which checks if we should hide the Languages page. This
  checks if the EFL feature is active, and if one of the following is
  true:
  * The application_form is unsubmitted
  * The application_form#english_main_language field is nil (ie - it's
    never been completed)
- Use this object to:
  * Skip the Languages page when filling in Personal Details
  * Hide the corresponding Review rows.
  * Allow the section to be marked complete even if the Languages data
    is missing.