Update TestApplications to treat a minimal set of courses as 'full'
Introduce a 'course_full' option to #create_application. If this is
true, always use the first n courses_to_apply_to when creating
application choices.

This ensures we can reliably create applications to full courses while
filling vacancies for the least number of courses possible. It improves
upon the previous implementation of GenerateTestApplications by removing
the possibility of gradually filling all available courses with
multiple sequential runs of the task.