Only show providers organisations they can manage
The manage-orgs view used to show all providers on either side of any
relationship with the current provider, whether or not the current user
had manage_orgs permissions for them.

The provider user will be able to see all the relevant relationships on
the show pages for each of their providers, so these extra entries in
the organisations#index page were just (slightly misleading) noise.