Centralise query to figure out providers that the user has manage_user permissions for
This query exists in some form in 3 places. This centralises it into
the `ProviderAuthorisation` class, because that's the place that should
know about permission-related things.

Also removes the `manage_users` scope because I think it's clearer to
read `provider_permissions.where(manage_users: true)` instead of
`provider_permissions.manage_users`.