Add JavaScript to the personal details nationalities page
This should do a few things. It should have an add
nationalities button. When clicked this should display
an addtional select for candidates to pick an additional
nationality. This button should not display once 3 selects
are visible.

It should have a remove link on the second and third select.
When clicked it should hide the select and clear it's value.