Update content when a provider can only view applications
We will be validating that at least one of the two providers in a relationship
will be authorised for each permission. If a provider has no authorisation for
any permission we display this as a sentence as there would only ever be one
provider for this case.

There is a case where the permissions have not yet been set up, in this
case we render a specific message explaining that the training provider
should set up the permissions.