Simplify changed_offer email
If the conditions of the offer are changed, we would send an email that is confusing:

```
Mallowpond University has changed the details of your offer.

Previous offer
Business (XVNK) at Main site

New offer
Business (XVNK) at Main site
```

Instead, we could not include details about the previous and new offer in the email.