Audit ProviderRelationshipPermissions
It is not possible to associate audits with multiple providers. We are
associating it with the training_provider as they are able to make changes. To
display the audit trail we decided to write a custom query to get the audits
for ratified provider permissions and add those to ratifying provider's
own_and_associated_audits in the next commit.