Fix referee replacement double render error
We had a Sentry error come in from a user that triggered a double render
error in the AdditionalRefereesController: https://sentry.io/organizations/dfe-bat/issues/1790159249/?project=1765973

This is because we call the
`redirect_to_confirm_if_no_more_reference_needed` method, and return
from its context, but don't return in the caller method.

This can lead to a `redirect_to` being run, with a follow-up `render`
call later down.

The solution is to `return if` the method returns something truthy.
The method returns nil or the value of `redirect_to`, which works well
enough.