Simplify LanguagesSectionPolicy boolean check
The existing implementation checks if the application form is
pre-submission. Remove this check, such that `hide?` is true only when:

- The efl_section feature flag is active
- The english_main_language field is nil (ie - the candidate has never
  submitted values for the Languages page)

This solves a potential issue where a candidate who has filled in this
part of the form, but not yet submitted, can no longer edit their
answers. The simpler boolean check is sufficient to ensure we don't show
the page to anyone who doesn't need to see it.