Display ProviderRelationshipPermissions audits for ratifying provider
We want to display changes to ProviderRelationshipPermissions for both sides of
the relationship training providers and ratifying provider.
ProviderRelationshipPermissions audits are associated only with the training
provider using the audited gem. It is currently not possible to have more than
one associations.

To display the audit trail we decided to write a custom query to get the audits
for ratified provider permissions and add those to ratifying provider's
own_and_associated_audits.