Override language-related fields on ApplicationForm
We've introduced an 'English as a foreign language' section to the
application. It appears if a candidate selects a nationality other than
British or Irish. Switching on the feature also hides the Languages
section of Personal Details if the candidate hasn't already filled this
in, as there is overlap between the two sets of questions and EFL is
intended as a replacement.

If the Languages section is hidden, the application can be submitted
with the following database fields containing null values:

- english_main_language
- english_language_details

Both are used in the Vendor API and in the Manage interface. To ensure
we continue to provide consistent data in these fields, we need to infer
values for them if they are blank:

- english_main_language
  * Return true if the candidate is British/Irish OR if they declare
    that they don't need an English qualification.
  * Return false otherwise.
- english_language_details
  * Return a formatted string description of the candidate's EFL
    qualification, if one has been provided.