Fix bug when hiding Languages page
The Languages page in Personal Details should be hidden if
LanguagesSectionPolicy#hide? returns true. This doesn't happen currently
if the candidate selects a non-British nationality and has to complete
the Right to Work page.

Update the Right to Work controller to ensure we invoke
LanguagesSectionPolicy when deciding where to redirect after form
submission.