Relax check for next provider needing permissions setup
We do not validate that permissions should be non-empty so it doesn't make sense
to include the check when calculating the next provider.

This commit attempts to rectify a bug where a user amending a set of permissions
in the wizard will be returned to the same permissions set unless they turn a permission
'on', this isn't a validation requirement so remove the check for empty permissions.