Fix Personal Details system spec
This spec checks the database value for
AppliationForm#english_main_language in one of its steps to make it
clear what the expected state is for subsequent assertions.

As part of deprecating the Languages section of the application, we now
override the `english_main_language` method. To check if the database
value of this field is nil (ie - the Languages section has never been
filled in), we need to use a different syntax.