Remove random nationalities from application_form factory
There is behaviour in the system which is dependent on the nationality
chosen (ie - the EFL section of the form only appears when the
nationality is something other than British or Irish). As such, it's
helpful to have a consistent, predictable value in the factory, with
tests themselves changing the nationality as required.