Fix calculation of RBD days remaining in app card component
This was calculating the number of days remaining based on the date
difference so could lead to a lower value than I think is intended, e.g.
if it's 23:59 and the RBD is 0:01 the following day we probably don't
want to say that you have '1 day remaining'. It's mostly a testing
concern because RBDs are normally at the end of the day so you would
normally get the correct result in real life. (there were test failures
when run between 23:00 and 0:00.)