Validate permissions fields
We need to validate permissions fields so that error messages can be replayed in the form against the correct fields.
This gets tricky because we store the entire set of permissions for the wizard in a hash format, so we need to convert the values from the fields into
the correct storage format as well as validate the values.