Add grade autocomplete to Maths and English GCSEs
Add autocomplete Javascript to the grade field when adding a Maths or
English GCSE.

We're leaving Science grades out for now, as that page currently only
has a single free-form input field. Supporting autocomplete for Triple
Science award grading in a single field isn't really viable. There's a
design for reworking the Science flow so that it has separate grade
fields for Triple award. We'll revisit autocomplete when implementing
that design.

In addition to introducing the autocomplete Javascript, this commit:

- uses a select box for the enhanced grade field, as GCSE grades are a
  known set of standardised values
- renders the existing free-form input field for all other qualification
  types, eg - Scottish National 5