Add EndOfCycleTimetable class
There are a number of dates we have to be aware of as we approach the
end of the current cycle, as various bits of functionality and messaging
will be affected them.

Introduce an EndOfCycleTimetable class to hold these dates. To begin
with, expose helper methods to check whether or not we should be showing
deadline messaging for Apply 1 and Apply 2 applications.