Fix AbstractController::DoubleRenderError in ProviderUsersController
find_provider_user was called directly in an action. We think this may have
resulted in double rendering as the action would not have halted at the first
render. So moved this to before action.  If a "before" filter renders or
redirects, the action will not run. If there are additional filters scheduled
to run after that filter, they are also cancelled.
[https://guides.rubyonrails.org/action_controller_overview.html#filters]

Co-authored-by: Steve Laing <steve.laing@gmail.com>