Fix back navigation issues.
When navigating back from the edit permissions form we should step back to the previous relationship.
When navigating back from the edit permissions form when we're amending a permissions set, we should return to the 'check' page.

This commit also removes the 'skip_further_permissions' wizard attribute
as the same functionality can be achieved with the 'checking_answers' attribute.