Add a `cycle` field to the application_forms table
Application forms don't contain any information about which cycle they
belong to. Add a database field to record cycle year.

For the convenience of this spike, the default is set to 2020. This
isn't something we'd do for real as anything other than a short-term
step to populate existing records. All newly-created application forms
would have to set the current cycle year.