Remove custom magic link Warden strategy
The MagicLinkStrategy implements custom authentication logic for use at
the Warden middleware layer. It tries to find a candidate by the token
in params, and authenticates them if the token is not expired.

The CandidateInterface::SignInController, however, customises our login
flow and invokes the Devise `sign_in` method directly. This in turn
invokes Warden to authenticate the correct candidate. The controller
also makes the same token expiry checks that MagicLinkStrategy does.

This results in an overlap of behaviour, such that by the time the
SignInController is calling `sign_in` with the candidate it finds, the
candidate is actually already signed in.

Essentially this means the MagicLinkStrategy isn't needed. Also, its
continued presence confuses attempts at understanding how our auth flow
works. Remove it, leaving the custom controller logic in place to handle
candidate authentication.