Add DSA success page with contextual content
This page acts as a mid-step when a user has accepted all DSAs and needs to then set up permissions for their organisation.
If the user does not have permissions to set up org permissions they see a link which can take them to their applications.
In this scenario they may also see content about managing users if they have the relevant permissions.