Check for expired token before confirming authentication
When clicking an expired magic link, the candidate is currently
presented with the sign in confirmation page and then informed that
their token is expired after they click 'Continue'.

Make a check for token expiry in the confirm_authentication action, so
that we take them straight to the expired token page.