GetApplicationFormsReadyToSendToProviders can handle cancelled choices
Because we were being paranoid about mixed statuses on application
forms, we didn't sent application_complete choices that shared forms
with cancelled choices.

The paranoia arose from the fact that references could conceivably be
out of sync across application choices, ie that one choice could have
been marked application_complete and the others still be
awaiting_references, which could mean in turn that we delivered the form
twice and emailed the candidate twice.

This caused us not to send forms which had a mixture of complete and
cancelled choices. Send those forms.