GetApplicationFormsReadyToSendToProviders can handle mixed choice states
We had a bug in production where an application form mixed up
application_complete and cancelled states on its choices. This caused
the completed choice never to be sent as
GetApplicationFormsReadyToSendToProviders did not consider those forms
ready to be sent. We ought not to pass SendApplicationToProvider choices
we know it can't handle - this means we can have it fail noisily in
future.