Be strict about how clients can call TextOrdinalizer
Hide its lookup table so clients aren’t tempted to do its business for
it.

We had a bug where the LinksToPreviousApplicationsComponent blew up
because it (legitimately) tried to ordinalize a number > 10 without
going through the proper interface, and ended up with an unexpected nil.

This change does not damage LinksToPreviousApplicationsComponent because
the fallback to Rails's #ordinalize method will return eg `11th` which
noops when you try to #capitalize it.