Add feature flag to simulate EoC cancellation behaviour
For a specific time window towards the end of the cycle, we cancel any
applications to unavailable course options before they can be sent to
the provider.

When this feature flag is active, the time-related checks for whether or
not we apply that behaviour are stubbed. This makes manual testing
easier when outside of the time window.