Remove wizard dep from ProviderRelationshipPermissionsListComponent
This component used to deal with a permissions model and a wizard that
might have some data to put in that model. Now that the wizard can
pass us populated (but unpersisted) models we can make the interface
more generic and this component becomes reuseable.

The spec only tested interaction with the wizard, so it's been removed
too.