Parse wizard permissions inside the wizard
SetupProviderRelationshipPermissions now knows nothing about the wizard:
the wizard is responsible for translating its own internal data
structure into ActiveRecord models.

Having this method on the wizard makes it possible to map the contents
of the wizard onto a set of ready-to-persist models, which allows us to
simplify the check answers page for the wizard. I think this also makes
the wizard a bit more cohesive.