Allow candidates to carry over from end of apply 2
Candidates can carry over an application and amend an application
between the end of apply 2 and the beginning of the next cycle. They
cannot add course choices until the next cycle opens.

Also renamed `EndOfCycleTimetable.next_cycle_opens` to `apply_reopens`
to make it a little clearer and more consistent with `find_reopens`.