Refactor EndOfCycleTimetable
This -hopefully- simplifies the time table code by introducing the
concept of a "schedule", which can be either real (the one we'll use on
production), or one of the simulated schedules (where we're currently
in mid-cycle, course-stop, or between cycles).

The idea is that we can make the schedule a single setting in the
database, instead of different feature flags (which are currently weird
because they're mutually exclusive - you can't simulate both mid-cycle
and between cycles at the same time).