Move ratified courses to a different page
These courses are quite different from a provider’s own courses. Having
them on their own page makes this clearer, and makes it clearer that
the buttons to sync and  open all courses only apply to courses that
are owned by the provider.

Adding some awful CSS so the navigation doesn’t wrap.