Conditionally hide Language Details on Provider interface
We've been hiding the Languages section of the application form since
introducing the English as a Foreign Language section. The only
candidates who see Languages are those that filled it in prior to EFL
going live.

When viewing an application choice as a Provider, hide the Language
Details section if this was never filled in by the candidate.