Render error page if candidate selects unavailable course
Unavailable means that there are no locations for which there are
vacancies and the site is still valid rather than just considering
vancancy status. This makes the flow more consistent with the filtering
of locations on the _Choose location_ page and should avoid ending up
with an empty list of locations to choose from.