Remove blank/empty arrays from params
These are caused by the form builder now including a `hidden_field` by
default inside a `govuk_check_boxes_fieldset`, so even when nothing is
checked a blank string is submitted. With regular Rails models this is
automagically taken care of but when passing to form objects it causes
failures.