Fix CancelUnsubmittedApplicationsWorker
This worker was failing if run a second time, after having processed at
least one unsubmitted application on an earlier run. This is because the
condition in the query that prevented it picking up previously processed
applications had not been updated since we changed the way the
processing itself worked. It should just check the status of the
associated application choices.