Update TabNavigationComponent elements and styles
Tabs have been redesigned in the design prototype, they resemble
the sub-navigation component but contain some enhancements, so
this commit adds the markup and styles from the design prototype.
Originally I attempted to merge existing tab markup and styles with
those from the design prototype. This became an override bunfight and it
seemed simpler to use the markup and styles from the prototype.