Change nested `span`s in summary lists to `dd`
Axe flagged A11y issues with the SummaryListComponent. The docs say that
only `dl` elements may only contain `dt` and `dd`. As we use `div`s to
markup rows in the summary lists this appeared to be a serious issue.
However further investigation implied that `div`s are permitted whereas
the `span` elements inside the `div`s are not. Solution seems pretty
simple but need to check that there are no unforeseen styling issues.