Clarify wizard/controller responsibilities; remove wizard steps
Remove all 'write' steps from the wizard (update_conditions, commit).
Steps are meant to be thought as as combinations of GET/POSTs for same symbol.
The controller is responsible for triggering validation/persistence.

The wizard is responsible for providing #next_step which the controller redirects_to
after a successful write. Hardcode all paths in the views instead of using next_path.