Add a rake task to create large numbers of test applications
We're testing some of the performance characteristics of the Apply
service in preparation for the larger number of candidates expected next
cycle. One of these tests involves seeing how the service handles
display and navigation across a much larger number of applications.

The rake task added in this commit provides a mechanism to create a
(candidate, application_form, application_choice) set multiple times.

- It accepts the id of an example ApplicationForm record to derive its
  test data from.
- It accepts a number specifying how many distinct applications should be
  created.
- Most of the work is delegated to the `BulkCreateTestApplications` class.

Inside `BulkCreateTestApplications`, we compose some SQL to create the
three records in a single query. The logic here endeavours to create a
minimal representation of an application in the system, and
intentionally leaves out some of the related data typically found on an
application (eg - english_proficiency, references, qualifications, etc).