Add data migration to update legacy degree grade values
When entering a degree grade, in addition to allowing free text input we
also offer the candidate a selection of radio buttons mapped to one of
several common values. We've provided radio buttons historically,
however the actual stored database value for these standard options
changed when we added structured HESA data to degrees. Rather than
saving a short string such as `upper_second`, we now simply persist the
HESA description (`Upper second-class honours (2:1)`).

This prompted a change to how we display these grades on the frontend. Instead
of translating database values such as 'upper_second' to a presentable
form via I18n, we now simply show the database value. This works fine
for anything with a standard HESA grade or a free-text input, but
doesn't look so great for older degree records that simply display
`upper_second` on the frontend.

We can easily map those older values to the current HESA descriptions.
This commit adds a migration which applies that mapping and also adds
the corresponding HESA code to the degree record.